streamlit/requirements.txt
streamlit==1.41.1
google-cloud-bigquery==3.27.0
db-dtypes==1.3.1
pandas==2.2.3
plotly==5.24.1
python-dotenv==1.0.1
pyarrow==18.1.0
